Update and rename FibonacciUpto47thTerm.cpp to FibonacciUpto93rdTerm.cpp

+++ FibonacciUpto93rdTerm.cpp
@@ -1,9 +1,9 @@
 #include<bits/stdc++.h>
 using namespace std;
 main(){
-	int n;
+	long long int n;
 	cin>>n;
-	int n1=0,n2=1,sum=0;
+	long long int n1=0,n2=1,sum=0;
 	cout<<n1<<" "<<n2<<" ";
 	n=n-2;
 		while(n--){
